# Sourcemap performance improvements