--- a/Cargo.toml
+++ b/Cargo.toml
@@ -73,6 +73,7 @@ log = "0.4.14"
 maplit = "1.0.2"
 notify = "4.0.17"
 opener = "0.5.0"
+rayon = "1.7.0"
 reqwest = { version = "0.11.10", features = ["blocking", "json", "native-tls-vendored"] }
 ritz = "0.1.0"
 roblox_install = "1.0.0"