@filiptibell 2023-04-28
> Can you add this to the [Changelog](https://github.com/rojo-rbx/rojo/blob/master/CHANGELOG.md)?

Sure thing, I tried to find a similar previous note in the changelog and https://github.com/rojo-rbx/rojo/pull/548 was the most similar so its more or less copy pasted from that.

> I'm not sure what impact pulling `rayon` as a dependency will have on build time or executable size

It's a pretty lightweight dependency afaik but pulls in some heavier things such as `crossbeam`. Lock file only shows `rayon` lib itself as a new dependency though, so seems we already depend on the rest.

---

I also went ahead and used `num_cpus` to cap the number of threads we create in the threadpool based on available cpu count so we don't create unnecessary threads for low-powered systems. This was another transitive dependency we already used so it won't increase build time or binary size.